Add files via upload
This is my first project on python. It is basically a number guessing game where you will be given 5 chances. It will ask you to enter a number and if your number is matched correctly with the generated number then you will win the game. But if your all the guessed numbers are wrong than you will lose the game.
Hope you all will like my game :)